pyproject.toml
[project]
name = "mini-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "psycopg2-binary>=2.9.10",
]
